[pickers][internal] Move own internals from lab internals to dedicated file (#25498)
