+++ docs/pages/api-docs/mobile-time-picker.json
@@ -24,7 +24,7 @@
     "disableOpenPicker": { "type": { "name": "bool" } },
     "getClockLabelText": {
       "type": { "name": "func" },
-      "default": "<TDate extends any>(\n  view: 'hours' | 'minutes' | 'seconds',\n  time: TDate,\n  adapter: MuiPickersAdapter<TDate>,\n) => `Select ${view}. Selected time is ${adapter.format(time, 'fullTime')}`"
+      "default": "<TDate extends any>(\n  view: ClockView,\n  time: TDate,\n  adapter: MuiPickersAdapter<TDate>,\n) => `Select ${view}. Selected time is ${adapter.format(time, 'fullTime')}`"
     },
     "getOpenDialogAriaText": {
       "type": { "name": "func" },
