+++ packages/material-ui-lab/src/DateTimePicker/shared.ts
@@ -1,0 +1,1 @@
+export type DateTimePickerView = 'year' | 'date' | 'month' | 'hours' | 'minutes';
