+++ packages/material-ui-lab/src/DayPicker/PickersCalendarHeader.tsx
@@ -17,7 +17,7 @@ import {
   usePreviousMonthDisabled,
   useNextMonthDisabled,
 } from '../internal/pickers/hooks/date-helpers-hooks';
-import { DatePickerView } from '../internal/pickers/typings/Views';
+import { DayPickerView } from './shared';
 
 export type ExportedCalendarHeaderProps<TDate> = Pick<
   PickersCalendarHeaderProps<TDate>,
@@ -48,15 +48,15 @@ export interface PickersCalendarHeaderProps<TDate>
     switchViewButton?: any;
   };
   currentMonth: TDate;
-  views: DatePickerView[];
+  views: DayPickerView[];
   /**
    * Get aria-label text for switching between views button.
    */
-  getViewSwitchingButtonText?: (currentView: DatePickerView) => string;
+  getViewSwitchingButtonText?: (currentView: DayPickerView) => string;
   onMonthChange: (date: TDate, slideDirection: SlideDirection) => void;
-  openView: DatePickerView;
+  openView: DayPickerView;
   reduceAnimations: boolean;
-  onViewChange?: (view: DatePickerView) => void;
+  onViewChange?: (view: DayPickerView) => void;
 }
 
 export type PickersCalendarHeaderClassKey =
@@ -107,7 +107,7 @@ export const styles: MuiStyles<PickersCalendarHeaderClassKey> = (
   },
 });
 
-function getSwitchingViewAriaText(view: DatePickerView) {
+function getSwitchingViewAriaText(view: DayPickerView) {
   return view === 'year'
     ? 'year view is open, switch to calendar view'
     : 'calendar view is open, switch to year view';
