+++ packages/material-ui-lab/src/DayPicker/index.ts
@@ -2,3 +2,4 @@ export { default } from './DayPicker';
 
 export type DayPickerClassKey = import('./DayPicker').DayPickerClassKey;
 export type DayPickerProps<TDate> = import('./DayPicker').DayPickerProps<TDate>;
+export type DayPickerView = NonNullable<DayPickerProps<unknown>['view']>;
