--- packages/material-ui-lab/src/internal/pickers/constants/ClockType.ts
@@ -1,1 +1,0 @@
-export type ClockViewType = 'hours' | 'minutes' | 'seconds';
