+++ packages/material-ui-lab/src/internal/pickers/constants/dimensions.ts
@@ -1,5 +1,3 @@
-export const CLOCK_WIDTH = 220;
-export const CLOCK_HOUR_WIDTH = 36;
 export const DAY_SIZE = 36;
 export const DAY_MARGIN = 2;
 export const DIALOG_WIDTH = 320;
