+++ packages/material-ui-lab/src/internal/pickers/typings/Views.ts
@@ -1,7 +1,1 @@
 export type AllAvailableViews = 'year' | 'date' | 'month' | 'hours' | 'minutes' | 'seconds';
-
-export type DateTimePickerView = 'year' | 'date' | 'month' | 'hours' | 'minutes';
-
-export type DatePickerView = 'year' | 'date' | 'month';
-
-export type TimePickerView = 'hours' | 'minutes' | 'seconds';
