+++ packages/material-ui-lab/src/internal/pickers/typings/helpers.ts
@@ -1,5 +1,3 @@
-import { WithStyles } from '@material-ui/core/styles';
-
 /**
  * All standard components exposed by `material-ui` are `StyledComponents` with
  * certain `classes`, on which one can also set a top-level `className` and inline
@@ -14,12 +12,3 @@ export type MakeOptional<T, K extends keyof T> = {
   [P in K]?: T[P] | undefined;
 } &
   Omit<T, K>;
-
-export type MakeRequired<T, K extends keyof T> = {
-  [X in Exclude<keyof T, K>]?: T[X];
-} &
-  {
-    [P in K]-?: T[P];
-  };
-
-export type WithoutClasses<TProps extends WithStyles<any>> = Omit<TProps, keyof WithStyles<any>>;
